external consumer - attempt to fix travis tests failures 4

+++ qa/common/src/main/java/org/eclipse/kapua/qa/common/BasicSteps.java
@@ -60,13 +60,13 @@ public class BasicSteps extends TestBase {
         beforeCommon(scenario);
     }
 
-    @Before(value="@env_embedded_minimal", order=0)
+    @Before(value="@env_embedded_minimal and @setup", order=0)
     public void beforeScenarioEmbeddedMinimal(Scenario scenario) {
         beforeCommon(scenario);
         beforeNoDocker();
     }
 
-    @Before(value="@env_none", order=0)
+    @Before(value="@env_none and @setup", order=0)
     public void beforeScenarioNone(Scenario scenario) {
         beforeCommon(scenario);
         beforeNoDocker();
@@ -77,12 +77,12 @@ public class BasicSteps extends TestBase {
         afterScenarioDocker(scenario);
     }
 
-    @After(value="@env_embedded_minimal", order=0)
+    @After(value="@env_embedded_minimal and not (@setup and @teardown)", order=0)
     public void afterScenarioEmbeddedMinimal(Scenario scenario) {
         afterScenarioNoDocker(scenario);
     }
 
-    @After(value="@env_none", order=0)
+    @After(value="@env_none and not (@setup and @teardown)", order=0)
     public void afterScenarioNone(Scenario scenario) {
         afterScenarioNoDocker(scenario);
     }
@@ -102,23 +102,27 @@ public class BasicSteps extends TestBase {
     }
 
     protected void afterScenarioDocker(Scenario scenario) {
-        logger.info("Database cleanup...");
-        database.deleteAll();
-        logger.info("Database cleanup... DONE");
-        SecurityUtils.getSubject().logout();
-        KapuaSecurityUtils.clearSession();
+        if (!shutdown) {
+            logger.info("Database cleanup...");
+            database.deleteAll();
+            logger.info("Database cleanup... DONE");
+            SecurityUtils.getSubject().logout();
+            KapuaSecurityUtils.clearSession();
+        }
     }
 
     protected void afterScenarioNoDocker(Scenario scenario) {
-        logger.info("Database drop...");
-        try {
-            database.dropAll();
-            logger.info("Database drop... DONE");
-            database.close();
-        } catch (Exception e) {
-            logger.error("Failed execute @After", e);
+        if (!shutdown) {
+            logger.info("Database drop...");
+            try {
+                database.dropAll();
+                logger.info("Database drop... DONE");
+                database.close();
+            } catch (Exception e) {
+                logger.error("Failed execute @After", e);
+            }
+            KapuaSecurityUtils.clearSession();
         }
-        KapuaSecurityUtils.clearSession();
     }
 
     @When("Set test shutdown")
